Удалены строки с Debug.Log

+++ Assets/Game/Scripts/Domain/GameEntities/Core/Health/HealthUseCase.cs
@@ -23,7 +23,7 @@ namespace SampleGame
 
         public static float GetPercent(in this Health health)
         {
-            return health.Current / health.Max;
+            return health.Current / (float)health.Max;
         }
     }
 }
